+++ chat/models/__init__.py
@@ -1,0 +1,2 @@
+from .lobby import Lobby as _Lobby
+from .message import Message as _Message