+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to Porsche backend project will be documented in this file.
 
 ## Unreleased
+### Added
+- Users can find, create and delete chat rooms
 
 ##[0.1.0]() - 2021-05-19
 ### Added